#4484 python int type hint was getting translated to c int
what we want is unsigned long, so have to use 'long' which gets translated correctly to:
__Pyx_PyInt_From_unsigned_long(__pyx_v_xevent.xproperty.time);

+++ xpra/x11/bindings/window.pyx
@@ -1431,7 +1431,7 @@ cdef class X11WindowBindingsInstance(X11CoreBindingsInstance):
                                   xwindow, cursor, CurrentTime)
         return r == 0
 
-    def get_server_time(self, Window xwindow) -> int:
+    def get_server_time(self, Window xwindow) -> long:
         cdef unsigned char c = b"a"
         cdef Atom timestamp_prop = self.str_to_atom("XPRA_TIMESTAMP_PROP")
         XChangeProperty(self.display, xwindow, timestamp_prop,
